requirements.txt
fastapi == 0.111.0
uvicorn == 0.30.1
azure-monitor-opentelemetry == 1.5.0
azure-monitor-events-extension == 0.1.0